ecommerce-microservices/services/user-service/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
psycopg2-binary==2.9.9
pydantic[email]==2.5.3
python-jose[cryptography]==3.3.0
PyJWT==2.8.0
